Expose public API
Allow to use the library as a CLI tool as well as programmatically:
* Move all files to lib (instead src)
* Move cli tool to cli.js
* Expose public functions from index.js
