+++ lib/amend-group-stats.js
