+++ lib/build-output-summary.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const fs = require('fs');
+const path = require('path');
 const processData = require('./process-data');
 
 function rank(entries) {
@@ -9,10 +10,10 @@ function rank(entries) {
   return entries;
 }
 
-module.exports = function outputSummary(path) {
-  let entries = fs.readdirSync(path).
+module.exports = function outputSummary(statsPath) {
+  let entries = fs.readdirSync(statsPath).
     filter(file => /\.out\.json$/.test(file)).
-    map(file => JSON.parse(fs.readFileSync(path + file, 'UTF8'))).
+    map(file => JSON.parse(fs.readFileSync(path.join(statsPath, file), 'UTF8'))).
     map(processData);
 
   return {
