+++ lib/get-total.js
