+++ lib/group-files.js
@@ -54,4 +54,4 @@ module.exports = function groupFiles(files) {
   });
 
   return groups;
-}
+};
