+++ lib/index.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const buildOutputSummary = require('./build-output-summary');
+const summarize = require('./summarize');
+const summarizeAll = require('./summarize-all');
+
+module.exports = {
+  buildOutputSummary,
+  summarize,
+  summarizeAll
+};
